Fixed bug where some fields were null in dto and moved class to separate file.
