+++ search/src/main/java/com/jacob/wakatimeapp/search/data/network/mappers/ProjectDetailsDTOMapper.kt
@@ -1,8 +1,7 @@
-@file:Suppress("Filename") // ktlint-disable filename
-
-package com.jacob.wakatimeapp.search.data.network.mappers // ktlint-disable filename
+package com.jacob.wakatimeapp.search.data.network.mappers
 
 import com.jacob.wakatimeapp.core.common.data.dtos.ProjectDetailsDTO
+import com.jacob.wakatimeapp.search.domain.models.ProjectDetails
 import com.jacob.wakatimeapp.search.domain.models.ProjectListWithPageNumber
 
 fun ProjectDetailsDTO.toModel() = data.map { ProjectDetails(name = it.name) }
@@ -11,8 +10,4 @@ fun ProjectDetailsDTO.toModelWithPageNumber() = ProjectListWithPageNumber(
     projectList = this.toModel(),
     pageNumber = page,
     totalPageCount = totalPages,
-)
-
-data class ProjectDetails(
-    val name: String,
 )
