+++ search/src/main/java/com/jacob/wakatimeapp/search/domain/models/ProjectDetails.kt
@@ -1,0 +1,5 @@
+package com.jacob.wakatimeapp.search.domain.models
+
+data class ProjectDetails(
+    val name: String,
+)
