+++ search/src/main/java/com/jacob/wakatimeapp/search/domain/models/ProjectListWithPageNumber.kt
@@ -1,7 +1,5 @@
 package com.jacob.wakatimeapp.search.domain.models
 
-import com.jacob.wakatimeapp.search.data.network.mappers.ProjectDetails
-
 data class ProjectListWithPageNumber(
     val projectList: List<ProjectDetails>,
     val pageNumber: Int,
